+++ src/Hooks/CommandStartingListener.php
@@ -52,7 +52,7 @@ final class CommandStartingListener
         try {
             match ($event->command) {
                 'queue:work', 'queue:listen', 'horizon:work', 'vapor:work' => $this->registerJobHooks($event),
-                'schedule:run', 'schedule:work' => $this->registerScheduledTaskHooks(),
+                'schedule:run', 'schedule:work', 'vapor:schedule' => $this->registerScheduledTaskHooks(),
                 'help', 'inspire', 'schedule:finish' => null,
                 default => $this->registerCommandHooks($event),
             };